# Make a number of performance optimizations