… internal/zinc-persist/src/main/scala/sbt/internal/inc/binary/converters/ProtobufWriters.scala