I noticed that running https://github.com/eatkins/scala-build-watch-performance with a local sbt build using zinc 1.4.0-M1 was slower than 1.3.10 in turbo mode. Using flamegraphs, I was able to identify to hotspots and optimize them with some relatively straightforward changes. Before these changes, I got the following results with sbt 1.4.0-SNAPSHOT:
 project | min (ms) | max (ms) | median (ms) | total (ms) | cpu % |
:------- | :------: | :------: | :-------: | :--------: | :---: |
sbt-1.4.0-SNAPSHOT (3 source files) | 151 | 263 | 166 | 34787 | 0.2
sbt-1.4.0-SNAPSHOT (5003 source files) | 747 | 1016 | 836 | 103423 | 0.1

After these changes, the performance improved:
 project | min (ms) | max (ms) | median (ms) | total (ms) | cpu % |
:------- | :------: | :------: | :-------: | :--------: | :---: |
sbt-1.4.0-SNAPSHOT (3 source files) | 140 | 346 | 149 | 27990 | 0.2
sbt-1.4.0-SNAPSHOT (5003 source files) | 458 | 574 | 470 | 68228 | 0.2

Note that there is a lot of time spent stamping source files outside of zinc, scalatest and the scala compiler so the zinc overhead was probably about 400ms before these optimizations and 100ms after. 

By comparison, for sbt 1.3.10, I got
 project | min (ms) | max (ms) | median (ms) | total (ms) | cpu % |
:------- | :------: | :------: | :-------: | :--------: | :---: |
sbt-1.3.10-turbo (3 source files) | 119 | 220 | 125 | 25906 | 3.1
sbt-1.3.10-turbo (5003 source files) | 535 | 746 | 558 | 77775 | 3.5

I couldn't quite figure out why sbt 1.4.0-SNAPSHOT was consistently 20ms slower than 1.3.10 in the 3 source file example. It was reproducible in the shell though. I ran sbt with `-Dsbt.task.timings=true` and it appeared that about 10ms of the extra time was in `compileIncremental` and the other 10ms was somewhere that didn't show up in the task timing report.

The variance seemed generally lower after these changes which is probably because I reworked things so that less garbage is generated so GC is probably running less often.

I also just benchmarked running `compile` in the sbt shell in a scenario when we just needed to compile a single source file (`object Foo`) and with the head of origin/develop in zinc, it took about 700ms when there were 5000 other source files and with these zinc changes, it dropped down to 500ms.