@retronym 2020-07-20
LGTM. I have a few WIP performance improvements but this looks like a great start and I'll rebase atop it!